bunch of minor updates
Changed menu background colour, text size and name

Changed text box so the font now scales with the window and now wraps
around the text box.

Updated processing window function. Added a name at the top of the
window.

Edited the update text function so it now takes the player to the menu
when the text is finished rather than repeating.
Update txt file accordingly.